+++ Sources/App/Settings/Widgets/TileCard.swift
@@ -32,9 +32,9 @@ struct TileCard: View {
         .frame(maxWidth: .infinity)
         .frame(height: 55)
         .background(Color.asset(Asset.Colors.tileBackground))
-        .clipShape(RoundedRectangle(cornerRadius: 10))
+        .clipShape(RoundedRectangle(cornerRadius: CornerRadiusSizes.oneAndMicro))
         .overlay {
-            RoundedRectangle(cornerRadius: 10)
+            RoundedRectangle(cornerRadius: CornerRadiusSizes.oneAndMicro)
                 .stroke(Color.asset(Asset.Colors.tileBorder), lineWidth: 1)
         }
     }
@@ -55,7 +55,7 @@ struct TileCard: View {
     }
     .frame(maxWidth: .infinity, alignment: .center)
     .background(Color.asset(Asset.Colors.primaryBackground))
-    .clipShape(RoundedRectangle(cornerRadius: 16))
+    .clipShape(RoundedRectangle(cornerRadius: CornerRadiusSizes.two))
     .shadow(color: .black.opacity(0.2), radius: 10)
     .padding()
 }
