+++ Sources/App/WebView/ConnectionErrorDetailsView.swift
@@ -43,7 +43,7 @@ struct ConnectionErrorDetailsView: View {
                                 .frame(maxWidth: 600)
                                 .padding()
                                 .background(Color(uiColor: .secondarySystemBackground))
-                                .clipShape(RoundedRectangle(cornerRadius: 12))
+                                .clipShape(RoundedRectangle(cornerRadius: CornerRadiusSizes.oneAndHalf))
                             }
 
                             if let server, server.info.connection.canUseCloud,
@@ -72,7 +72,7 @@ struct ConnectionErrorDetailsView: View {
                         .frame(maxWidth: 600)
                         .padding()
                         .background(Color(uiColor: .secondarySystemBackground))
-                        .clipShape(RoundedRectangle(cornerRadius: 12))
+                        .clipShape(RoundedRectangle(cornerRadius: CornerRadiusSizes.oneAndHalf))
                         .padding(.top)
 
                         Rectangle()
