+++ Sources/Shared/DesignSystem/Constants/CornerRadiusSizes.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+// 8-point system
+public enum CornerRadiusSizes {
+    public static var half: CGFloat = 4
+    public static var one: CGFloat = 8
+    public static var oneAndHalf: CGFloat = 12
+    public static var two: CGFloat = 16
+    public static var three: CGFloat = 24
+    public static var four: CGFloat = 32
+    public static var five: CGFloat = 40
+    public static var six: CGFloat = 48
+
+    // Out of system points
+    public static var micro: CGFloat = 2
+    public static var oneAndMicro: CGFloat = 10
+    public static var twoAndMicro: CGFloat = 18
+}
