+++ Sources/Shared/DesignSystem/Styles/HAButtonStyles.swift
@@ -1,6 +1,12 @@
 import Foundation
 import SwiftUI
 
+enum HAButtonStylesConstants {
+    static var height: CGFloat = 55
+    static var cornerRadius: CGFloat = 12
+    static var disabledOpacity: CGFloat = 0.5
+}
+
 public struct HAButtonStyle: ButtonStyle {
     @Environment(\.isEnabled) private var isEnabled: Bool
 
@@ -9,10 +15,10 @@ public struct HAButtonStyle: ButtonStyle {
             .font(.callout.bold())
             .foregroundColor(.white)
             .frame(maxWidth: .infinity)
-            .frame(height: 55)
+            .frame(height: HAButtonStylesConstants.height)
             .background(isEnabled ? Color.asset(Asset.Colors.haPrimary) : Color.gray)
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .opacity(isEnabled ? 1 : 0.5)
+            .clipShape(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius))
+            .opacity(isEnabled ? 1 : HAButtonStylesConstants.disabledOpacity)
     }
 }
 
@@ -24,10 +30,10 @@ public struct HANeutralButtonStyle: ButtonStyle {
             .font(.callout.bold())
             .foregroundColor(.white)
             .frame(maxWidth: .infinity)
-            .frame(height: 55)
+            .frame(height: HAButtonStylesConstants.height)
             .background(Color.gray)
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .opacity(isEnabled ? 1 : 0.5)
+            .clipShape(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius))
+            .opacity(isEnabled ? 1 : HAButtonStylesConstants.disabledOpacity)
     }
 }
 
@@ -39,10 +45,10 @@ public struct HANegativeButtonStyle: ButtonStyle {
             .font(.callout.bold())
             .foregroundColor(.white)
             .frame(maxWidth: .infinity)
-            .frame(height: 55)
+            .frame(height: HAButtonStylesConstants.height)
             .background(isEnabled ? .red : Color.gray)
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .opacity(isEnabled ? 1 : 0.5)
+            .clipShape(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius))
+            .opacity(isEnabled ? 1 : HAButtonStylesConstants.disabledOpacity)
     }
 }
 
@@ -54,9 +60,9 @@ public struct HASecondaryButtonStyle: ButtonStyle {
             .font(.callout.bold())
             .foregroundColor(Color.asset(Asset.Colors.haPrimary))
             .frame(maxWidth: .infinity)
-            .frame(height: 55)
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .opacity(isEnabled ? 1 : 0.5)
+            .frame(height: HAButtonStylesConstants.height)
+            .clipShape(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius))
+            .opacity(isEnabled ? 1 : HAButtonStylesConstants.disabledOpacity)
     }
 }
 
@@ -68,9 +74,9 @@ public struct HASecondaryNegativeButtonStyle: ButtonStyle {
             .font(.callout.bold())
             .foregroundColor(.red)
             .frame(maxWidth: .infinity)
-            .frame(height: 55)
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .opacity(isEnabled ? 1 : 0.5)
+            .frame(height: HAButtonStylesConstants.height)
+            .clipShape(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius))
+            .opacity(isEnabled ? 1 : HAButtonStylesConstants.disabledOpacity)
     }
 }
 
@@ -94,11 +100,14 @@ public struct HACriticalButtonStyle: ButtonStyle {
             .font(.callout.bold())
             .foregroundColor(.black)
             .frame(maxWidth: .infinity)
-            .frame(height: 55)
+            .frame(height: HAButtonStylesConstants.height)
             .padding()
             .background(.red.opacity(0.5))
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .overlay(RoundedRectangle(cornerRadius: 12).stroke(Color.red, lineWidth: 1))
+            .clipShape(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius))
+            .overlay(RoundedRectangle(cornerRadius: HAButtonStylesConstants.cornerRadius).stroke(
+                Color.red,
+                lineWidth: 1
+            ))
     }
 }
